Shifted all the interface types in Chatbox to src/types/Chat.ts
