+++ src/components/chat/w2wChat/messageFeed/MessageFeed.tsx
@@ -8,7 +8,8 @@ import styled, { useTheme } from 'styled-components';
 
 // Internal Components
 import { useWeb3React } from '@web3-react/core';
-import { Feeds, User } from 'api';
+// import { Feeds, User } from 'api';
+import { Feeds, User } from 'types/Chat';
 import ChatSnap from 'components/chat/chatsnap/ChatSnap';
 import LoaderSpinner, { LOADER_TYPE } from 'components/reusables/loaders/LoaderSpinner';
 import { ItemVV2, SpanV2 } from 'components/reusables/SharedStylingV2';
@@ -24,6 +25,7 @@ import './MessageFeed.css';
 
 // Internal Configs
 import GLOBALS from 'config/Globals';
+
 
 
 interface MessageFeedProps {
