+++ src/helpers/w2w/user.ts
@@ -1,4 +1,6 @@
-import { ConnectedUser } from "api";
+// import { ConnectedUser } from "api";
+
+import { ConnectedUser } from "types/Chat";
 
 export function checkConnectedUser(connectedUser:ConnectedUser): boolean {
   if (
